Finishing up setup activity

+++ app/src/main/java/tendersaucer/slimm/CalorieUtils.java
@@ -5,7 +5,7 @@ package tendersaucer.slimm;
  */
 public final class CalorieUtils {
 
-    public static int getBMR(boolean isMale, int ageYears, float weightKg, int heightCm) {
+    public static int getBMR(boolean isMale, int ageYears, float weightKg, float heightCm) {
         int correction = isMale ? 5 : -161;
         return Math.round((10 * weightKg) + (6.25f * heightCm) - (5 * ageYears) + correction);
     }
